lab3/Lab3_Doroshenko_Herashchenko_Voloshyn/requirements.txt
pycryptodome>=3.17.0